pyproject.toml
[project]
name = "projet-machine-learning-no2"
version = "0.1.0"
description = "API de prédiction de churn client avec Machine Learning"
authors = [
    {name = "Wilfred"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pandas (>=2,<3)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "scipy (>=1.17.0,<2.0.0)",
    "plotly (>=6.5.2,<7.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "imblearn (>=0.0,<0.1)",
    "xgboost (>=3.1.3,<4.0.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
    "catboost (>=1.2.8,<2.0.0)",
    "fastapi (>=0.128.0,<0.129.0)",
    "unicorn (>=2.1.4,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "httpx",
    "ruff",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

# ============================================
# Configuration Ruff (Linter + Formatter)
# ============================================
[tool.ruff]
target-version = "py312"
line-length = 120
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "notebooks",
    "*.ipynb",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "W191",   # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B", "C4"]
"src/data_preprocessing.py" = ["B023"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ============================================
# Configuration Pytest
# ============================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

# ============================================
# Configuration Coverage
# ============================================
[tool.coverage.run]
source = ["api", "src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
